hi -
cppcheck... gives an internal error from toLongNumber for this source:

{{{
bool initialize(bool flag, float x1, float x2) {
  if (flag) {   }

  if (x1>-998. and x2>-998.) {
    return false;
  }

  return true;
}
}}}


{{{
$  cppcheck  x.cc
Checking x.cc ...
x.cc:0:0: error: Internal Error. MathLib::toLongNumber: input was not completely consumed: -998.and [cppcheckError]

^
}}}

+++ externals/simplecpp/simplecpp.cpp
@@ -929,7 +929,7 @@ void simplecpp::TokenList::combineOperators()
             if (tok->previous && tok->previous->number) {
                 tok->setstr(tok->previous->str() + '.');
                 deleteToken(tok->previous);
-                if (isFloatSuffix(tok->next) || (tok->next && tok->next->startsWithOneOf("AaBbCcDdEeFfPp"))) {
+                if (isFloatSuffix(tok->next) || (tok->next && tok->next->startsWithOneOf("AaBbCcDdEeFfPp") && tok->next->str() != "and")) {
                     tok->setstr(tok->str() + tok->next->str());
                     deleteToken(tok->next);
                 }
